+++ saint-coinach/src/main/java/ffxiv/housim/saintcoinach/io/IndexFile.java
@@ -39,7 +39,7 @@ public class IndexFile implements IIndexFile {
 
     @Override
     public int hashCode() {
-        return (int)(((datFile << 24) | packId.hashCode()) ^ offset);
+        return ((datFile << 24) | packId.hashCode()) ^ offset;
     }
 
     @Override
